requirements.txt
streamlit
tensorflow
opencv-python
Pillow
numpy
tqdm
requests
